refactor: separate model loader form dispatcher
