+++ src/index.ts
@@ -1,1 +1,1 @@
-export { default as Trunk } from "./core";
+export { default as Trunk } from './dispatcher';
